+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.3.0 (WIP)
+## 0.3.0 (2022-05-31)
 
 * Allow adding categories to components
 * Component settings update fixes